+++ examples/02_typescript/src/Counter.tsx
@@ -2,11 +2,9 @@ import React from 'react';
 
 import { useDispatch, useTrackedState } from 'reactive-react-redux';
 
-import { Action, State } from './state';
-
 const Counter = () => {
-  const state = useTrackedState<State>();
-  const dispatch = useDispatch<Action>();
+  const state = useTrackedState();
+  const dispatch = useDispatch();
   return (
     <div>
       {Math.random()}